pyproject.toml
[project]
name = "capstone-data-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "census>=0.8.25",
    "dbt-core>=1.11.6",
    "dbt-databricks>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "fredapi>=0.5.2",
    "pandas>=2.2.3",
    "prefect>=3.6.24",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
